feat: 가입하기 클릭 시 회원가입 페이지로 이동
